app/rag_api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.6
requests==2.32.3
psycopg2-binary==2.9.9
prometheus-client==0.21.0
starlette==0.41.3
python-dotenv
sentence-transformers==3.3.1
numpy==1.26.4
